Added add owner page, home page, menu
