+++ app/src/main/AndroidManifest.xml
@@ -22,11 +22,16 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MapsActivity"/>
+        <activity android:name=".MapsActivity"
+            android:parentActivityName=".HomeScreen"/>
+        <activity android:name=".HomeScreen"
+            android:parentActivityName=".MainActivity"/>
         <activity android:name=".RegisterUser"
-            android:windowSoftInputMode="adjustPan"/>
+            android:windowSoftInputMode="adjustPan"
+            android:parentActivityName="com.example.livestock.MainActivity"/>
         <activity android:name=".AddOwner"
-            android:windowSoftInputMode="adjustPan"/>
+            android:windowSoftInputMode="adjustPan"
+            android:parentActivityName="com.example.livestock.HomeScreen"/>
     </application>
 
 </manifest>
