+++ app/src/main/java/com/example/livestock/AddOwner.java
@@ -42,7 +42,8 @@ public class AddOwner extends Activity {
             public void onResponse(String response)
             {   // Get Response
 
-                if(response.contains("ongitude")){
+                if(response.contains("Owner added")){
+                    Toast.makeText(getApplicationContext(), response.replace("\"", ""), Toast.LENGTH_LONG).show();
                     finish();
                 }
                 else {
