+++ app/src/main/java/com/example/livestock/LivestockUtils.java
@@ -1,0 +1,4 @@
+package com.example.livestock;
+
+public class LivestockUtils {
+}
