+++ app/src/main/res/menu/menu_main.xml
@@ -1,0 +1,22 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.example.livestock.MainActivity">
+    <item
+        android:id="@+id/menu_login"
+        android:orderInCategory="99"
+        android:title="Login"
+        app:showAsAction="never">
+    </item>
+    <item
+        android:id="@+id/menu_add_owner"
+        android:orderInCategory="100"
+        android:title="Add Owner"
+        app:showAsAction="never" />
+    <item
+        android:id="@+id/menu_location_search"
+        android:orderInCategory="101"
+        android:title="Location Search"
+        app:showAsAction="never">
+    </item>
+</menu>
